Add build-all and clone-all scripts.
Added or1k-clone-all.sh script, and or1k-build-all.sh script.  The
intent is that the build-all script would be a little more flexible than
the existing build-elf script, with command line flags to control
different aspects of the build.

Right now, there's not that much control, the most interesting flag is
the --clean flag, without this you'll get a rebuild, saving a little
time if nothing has changed.

The or1k-clone-all.sh script provides an easy way to checkout the
repositories ready for the build-all script.

Both scripts make use of a common logging directory.

ChangeLog:

	* define-release.sh: New file.
	* or1k-build-all.sh: New file.
	* or1k-clone-all.sh: New file.